[Fix] Change FlightManager to Education Manager
