+++ Orum/Orum/ViewModels/FlightManager.swift
@@ -1,0 +1,16 @@
+//
+//  FlightManager.swift
+//  Orum
+//
+//  Created by 차차 on 10/31/23.
+//
+
+import Foundation
+
+class FlightManager: ObservableObject {
+    @Published var remainingTime: Int
+
+    init() {
+        remainingTime = 0
+    }
+}
